% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BulkPipeline.R
\name{config<-}
\alias{config<-}
\alias{config<-,FLAMES.Pipeline-method}
\title{Set pipeline configurations}
\usage{
config(pipeline) <- value

\S4method{config}{FLAMES.Pipeline}(pipeline) <- value
}
\arguments{
\item{pipeline}{An pipeline of class `FLAMES.Pipeline`.}

\item{value}{A list containing the configuration of the pipeline, or
a path to a JSON configuration file.}
}
\value{
An pipeline of class `FLAMES.Pipeline` with the updated configuration.
}
\description{
This function sets the configuration of the pipeline.
}
\examples{
pipeline <- example_pipeline(type = "BulkPipeline")
# Set a new configuration
config(pipeline) <- create_config(outdir = tempdir())
}
