% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BulkPipeline.R
\name{config}
\alias{config}
\alias{config,FLAMES.Pipeline-method}
\title{Get pipeline configurations}
\usage{
config(pipeline)

\S4method{config}{FLAMES.Pipeline}(pipeline)
}
\arguments{
\item{pipeline}{An object of class `FLAMES.Pipeline`.}
}
\value{
A list containing the configuration of the pipeline.
}
\description{
This function returns the configuration of the pipeline.
}
\examples{
pipeline <- example_pipeline(type = "BulkPipeline")
config(pipeline)
}
