% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BulkPipeline.R
\name{controllers}
\alias{controllers}
\alias{controllers,FLAMES.Pipeline-method}
\title{Get controllers}
\usage{
controllers(pipeline)

\S4method{controllers}{FLAMES.Pipeline}(pipeline)
}
\arguments{
\item{pipeline}{A FLAMES.Pipeline object.}
}
\value{
A named list of \code{crew_class_controller} objects, where each
controller corresponds to a step in the pipeline.
}
\description{
Gets the controllers for the pipeline.
}
\examples{
pipeline <- example_pipeline(type = "MultiSampleSCPipeline")
controllers(pipeline) # get the controllers
}
