% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_barcode.R
\name{.legacy_pattern_to_flexiplex_segments}
\alias{.legacy_pattern_to_flexiplex_segments}
\title{Convert legacy pattern to Flexiplex segments}
\usage{
.legacy_pattern_to_flexiplex_segments(
  pattern,
  barcodes_file,
  max_bc_editdistance,
  buffer_size
)
}
\arguments{
\item{pattern}{named character vector defining the barcode pattern}

\item{barcodes_file}{path to file containing barcode allow-list}

\item{max_bc_editdistance}{max edit distances for the barcode sequence}

\item{buffer_size}{buffer size for barcode matching}
}
\value{
a list of FlexiplexSegment objects
}
\description{
Convert legacy pattern to Flexiplex segments
}
\keyword{internal}
