% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_barcode.R
\name{.resolve_bc_lists}
\alias{.resolve_bc_lists}
\title{Resolve bc_list_name keys/indices to file paths}
\usage{
.resolve_bc_lists(x, barcodes_files)
}
\arguments{
\item{x}{list of FlexiplexSegment objects or list of FlexiplexGroup objects}

\item{barcodes_files}{named or unnamed (only allowd when length is 1) character vector of barcode file paths}
}
\description{
Resolve bc_list_name keys/indices to file paths
}
\keyword{internal}
