% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_long_pipeline.R
\name{example_pipeline}
\alias{example_pipeline}
\title{Example pipelins}
\usage{
example_pipeline(type = "SingleCellPipeline")
}
\arguments{
\item{type}{The type of pipeline to create. Options are "SingleCellPipeline",
"BulkPipeline", and "MultiSampleSCPipeline".}
}
\value{
A pipeline object of the specified type.
}
\description{
Provides example pipelines for bulk, single cell and multi-sample single cell.
}
\examples{
example_pipeline("SingleCellPipeline")

}
\seealso{
\code{\link{SingleCellPipeline}} for creating the single cell pipeline,
\code{\link{BulkPipeline}} for bulk long data,
\code{\link{MultiSampleSCPipeline}} for multi sample single cell pipelines.
}
