% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_isoform.R
\name{find_isoform}
\alias{find_isoform}
\title{Isoform identification}
\usage{
find_isoform(annotation, genome_fa, genome_bam, outdir, config)
}
\arguments{
\item{annotation}{Path to annotation file. If configured to use bambu, the annotation
must be provided as GTF file.}

\item{genome_fa}{The file path to genome fasta file.}

\item{genome_bam}{File path to BAM alignment file. Multiple files could be provided.}

\item{outdir}{The path to directory to store all output files.}

\item{config}{Parsed FLAMES configurations.}
}
\value{
The updated annotation and the transcriptome assembly will be saved in the
output folder as \code{isoform_annotated.gff3} (GTF if bambu is selected) and
\code{transcript_assembly.fa} respectively.
}
\description{
Long-read isoform identification with FLAMES or bambu.
}
\keyword{internal}
