% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAMES.R
\name{FLAMES}
\alias{FLAMES}
\title{FLAMES: full-length analysis of mutations and splicing}
\usage{
FLAMES()
}
\description{
FLAMES: full-length analysis of mutations and splicing
}
\examples{
basilisk::basiliskRun(env = FLAMES:::flames_env, fun = function(){})
}
