% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flexiplex}
\alias{flexiplex}
\title{Rcpp port of flexiplex}
\usage{
flexiplex(
  r_segments,
  r_barcode_groups,
  max_flank_editdistance,
  reads_in,
  reads_out,
  stats_out,
  bc_out,
  reverseCompliment,
  n_threads
)
}
\arguments{
\item{r_segments}{List defining the barcode structure}

\item{r_barcode_groups}{List defining barcode groups}

\item{max_flank_editdistance}{int, maximum edit distance for matching flanking sequences}

\item{reads_in}{Input FASTQ or FASTA file}

\item{reads_out}{output file for demultiplexed reads}

\item{stats_out}{output file for demultiplexed stats}

\item{bc_out}{WIP}

\item{reverseCompliment}{bool, whether to reverse complement the reads after demultiplexing}

\item{n_threads}{number of threads to be used during demultiplexing}
}
\value{
integer return value. 0 represents normal return.
}
\description{
demultiplex reads with flexiplex, for detailed description, see
documentation for the original flexiplex: https://davidsongroup.github.io/flexiplex
}
