% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_isoform.R
\name{get_GRangesList}
\alias{get_GRangesList}
\title{Parse FLAMES' GFF output}
\usage{
get_GRangesList(file)
}
\arguments{
\item{file}{the GFF file to parse}
}
\value{
A Genomic Ranges List
}
\description{
Parse FLAMES' GFF ouputs into a Genomic Ranges List
}
\keyword{internal}
