% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2_align.R
\name{gff2bed}
\alias{gff2bed}
\title{Convert GFF/GTF to BED file}
\usage{
gff2bed(gff, bed)
}
\arguments{
\item{gff}{Path to the GFF/GTF file}

\item{bed}{Path to the output BED file to be written}
}
\value{
invisible, the BED file is written to the specified path
}
\description{
Convert GFF/GTF to BED file
}
\keyword{internal}
