% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BulkPipeline.R
\name{plot_durations}
\alias{plot_durations}
\alias{plot_durations,FLAMES.Pipeline-method}
\title{Plot pipeline step durations}
\usage{
plot_durations(x)

\S4method{plot_durations}{FLAMES.Pipeline}(x)
}
\arguments{
\item{x}{A FLAMES.Pipeline object.}
}
\value{
A ggplot2 object.
}
\description{
This function creates a horizontal bar plot showing the duration
of each pipeline step using ggplot2.
}
\examples{
pipeline <- example_pipeline("BulkPipeline")
pipeline <- run_FLAMES(pipeline)
plot_durations(pipeline)
}
