% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BulkPipeline.R
\name{resume_FLAMES}
\alias{resume_FLAMES}
\alias{resume_FLAMES,FLAMES.Pipeline-method}
\title{Resume a FLAMES pipeline}
\usage{
resume_FLAMES(pipeline)

\S4method{resume_FLAMES}{FLAMES.Pipeline}(pipeline)
}
\arguments{
\item{pipeline}{A FLAMES.Pipeline object.}
}
\value{
An updated FLAMES.Pipeline object.
}
\description{
This function resumes a FLAMES pipeline by running configured
but unfinished steps.
}
\examples{
pipeline <- example_pipeline("BulkPipeline")
pipeline <- run_step(pipeline, "genome_alignment")
pipeline <- resume_FLAMES(pipeline)
}
\seealso{
\code{\link{run_FLAMES}} to run the entire pipeline.
}
