% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scmixology_lib10}
\alias{scmixology_lib10}
\title{scMixology short-read gene counts - sample 2}
\format{
## `scmixology_lib10`
A \code{SingleCellExperiment} with 7,240 rows and 60 columns:
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE154869>
}
\usage{
data(scmixology_lib10)
}
\value{
A \code{SingleCellExperiment} object
}
\description{
Short-read gene counts from long and short-read single 
cell RNA-seq profiling of human  lung adenocarcinoma cell 
lines using 10X version 2 chemstry. See
Tian, L. et al. Comprehensive characterization of single-cell
full-length isoforms in human and mouse with long-read sequencing.
Genome Biology 22, 310 (2021).
}
\keyword{datasets}
