% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scmixology_lib90}
\alias{scmixology_lib90}
\title{scMixology short-read gene counts - sample 1}
\format{
## `scmixology_lib90`
A \code{SingleCellExperiment}
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE126906>
}
\usage{
data(scmixology_lib90)
}
\value{
A \code{SingleCellExperiment} object
}
\description{
Short-read single cell RNA-seq profiling of 
human lung adenocarcinoma cell lines using 10X version 
2 chemstry. Single cells from five human lung adenocarcinoma
cell lines (H2228, H1975, A549, H838 and HCC827) were mixed 
in equal proportions and processed using the Chromium 10X platform, 
then sequenced using Illumina HiSeq 2500. See
Tian L, Dong X, Freytag S, Lê Cao KA et al. 
Benchmarking single cell RNA-sequencing analysis pipelines using
mixture control experiments. Nat Methods 2019 Jun;16(6):479-487. PMID: 31133762
}
\keyword{datasets}
