% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAMES_Pipeline.R
\name{show,FLAMES.Pipeline-method}
\alias{show,FLAMES.Pipeline-method}
\alias{show,FLAMES.SingleCellPipeline-method}
\alias{show,FLAMES.MultiSampleSCPipeline-method}
\title{Show method for FLAMES.Pipeline}
\usage{
\S4method{show}{FLAMES.Pipeline}(object)

\S4method{show}{FLAMES.SingleCellPipeline}(object)

\S4method{show}{FLAMES.MultiSampleSCPipeline}(object)
}
\arguments{
\item{object}{An object of class `FLAMES.Pipeline`}
}
\value{
None. Displays output to the console.
}
\description{
Displays the pipeline in a pretty format
}
\examples{
ppl <- example_pipeline()
show(ppl)
}
\keyword{internal}
