% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_long_pipeline.R
\name{experiment}
\alias{experiment}
\title{Get pipeline results}
\usage{
experiment(pipeline)
}
\arguments{
\item{pipeline}{A FLAMES.Pipeline object.}
}
\value{
A \code{SummarizedExperiment} object, a \code{SingleCellExperiment} object,
or a list of \code{SingleCellExperiment} objects.
}
\description{
This function returns the results of the pipeline as a
\code{SummarizedExperiment} object, a \code{SingleCellExperiment} object, or a
list of \code{SingleCellExperiment} objects, depending on the pipeline type.
}
\examples{
pipeline <- example_pipeline(type = "BulkPipeline")
pipeline <- run_FLAMES(pipeline)
se <- experiment(pipeline)
}
