% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_long_pipeline.R
\name{run_FLAMES}
\alias{run_FLAMES}
\title{Execute a FLAMES pipeline}
\usage{
run_FLAMES(pipeline)
}
\arguments{
\item{pipeline}{A FLAMES.Pipeline object.}
}
\value{
An updated FLAMES.Pipeline object.
}
\description{
This function runs the FLAMES pipeline. It will run all steps in the pipeline.
}
\examples{
pipeline <- example_pipeline("BulkPipeline")
pipeline <- run_FLAMES(pipeline)
}
\seealso{
\code{\link{resume_FLAMES}} to resume a pipeline from the last completed step.
}
